Bugfix - dont allow removal of plugins if asset is frozen
If an asset is frozen, no plugins should be removed until the asset is unfrozen. Use cases that are broken because of this functionality include non-custodial marketplaces and loans, as the holder can approve trransfer and freeze auth, the program can freeze the asset, then the user can remove the transfer auth, meaning the program is not able to transfer the asset.

* Updated freeze delegate check that was only denying the removal of the FreezeDelegate plugin
* Made the same change to PermanentFreezeDelegate

There are some other functions that should probably be denied if an asset is frozen as they result in unusual behaviour - eg you can add a freeze delegate and freeze an asset with permanent transfer auth, then transfer it while frozen. This should be addressed in a separate PR